chore: clear CI warnings
- bump `actions/setup-node` and `pnpm/action-setup` to the majors that run
  on Node 24, the runners now warn about the Node 20 based ones
- use the `simulation` runner mode in the example matrix, `instrumentation`
  is deprecated
- widen the root `engines.node` to `>=22`, the exact pin made every Node 22
  example job print an unsupported engine warning
- pass `sourceMap`/`outputToFilesystem` to the rollup config plugin, which
  otherwise warns while compiling `rollup.config.ts`
- move the runner mode helpers to their own module: `introspection.ts`
  imported them from the barrel file, so rollup reported a cycle
- ignore the `EVAL` warning for `optimization.ts`, %OptimizeFunctionOnNextCall
  has no non-eval entry point
- build the symbol names with `v8::String::Utf8Value`, `Utf8Length` and
  `WriteUtf8` are deprecated as of the V8 shipped with Node 24
