+++ .github/workflows/release.yml
@@ -28,8 +28,8 @@ jobs:
         with:
           fetch-depth: 0
           submodules: true
-      - uses: pnpm/action-setup@v2
-      - uses: actions/setup-node@v3
+      - uses: pnpm/action-setup@v4
+      - uses: actions/setup-node@v6
         with:
           cache: pnpm
           node-version-file: .nvmrc
@@ -51,8 +51,8 @@ jobs:
         with:
           fetch-depth: 0
           submodules: true
-      - uses: pnpm/action-setup@v2
-      - uses: actions/setup-node@v3
+      - uses: pnpm/action-setup@v4
+      - uses: actions/setup-node@v6
         with:
           cache: pnpm
           node-version-file: .nvmrc
