+++ package.json
@@ -48,6 +48,6 @@
     }
   },
   "engines": {
-    "node": "24.19.0"
+    "node": ">=22"
   }
 }
