+++ packages/core/package.json
@@ -16,7 +16,7 @@
   "types": "dist/index.d.ts",
   "gypfile": true,
   "scripts": {
-    "build": "NODE_NO_WARNINGS=1 rollup -c rollup.config.ts --configPlugin typescript --configImportAttributesKey with",
+    "build": "rollup -c",
     "build-native-addon": "prebuildify --name node --strip --no-napi --target 22.0.0 --target 24.0.0",
     "build-tracer-client": "openapi --client axios --input ./tracer.spec.json --name MongoTracer --output ./src/generated/openapi",
     "test": "jest --passWithNoTests --silent",
