+++ packages/core/src/index.ts
@@ -1,6 +1,7 @@
 import { checkV8Flags } from "./introspection";
 import { MongoMeasurement } from "./mongoMeasurement";
 import native_core from "./native_core";
+import { getCodspeedRunnerMode } from "./runnerMode";
 
 declare const __VERSION__: string;
 
@@ -9,44 +10,6 @@ const linuxPerf = new native_core.LinuxPerf();
 export const isBound = native_core.isBound;
 
 export const mongoMeasurement = new MongoMeasurement();
-
-type CodSpeedRunnerMode = "disabled" | "simulation" | "memory" | "walltime";
-
-type InstrumentMode = "disabled" | "analysis" | "walltime";
-
-export function getCodspeedRunnerMode(): CodSpeedRunnerMode {
-  const isCodSpeedEnabled = process.env.CODSPEED_ENV !== undefined;
-  if (!isCodSpeedEnabled) {
-    return "disabled";
-  }
-
-  // If CODSPEED_ENV is set, check CODSPEED_RUNNER_MODE
-  const codspeedRunnerMode = process.env.CODSPEED_RUNNER_MODE;
-  if (
-    codspeedRunnerMode === "instrumentation" ||
-    codspeedRunnerMode === "simulation"
-  ) {
-    return "simulation";
-  } else if (codspeedRunnerMode === "memory") {
-    return "memory";
-  } else if (codspeedRunnerMode === "walltime") {
-    return "walltime";
-  }
-
-  console.warn(
-    `Unknown codspeed runner mode: ${codspeedRunnerMode}, defaulting to disabled`,
-  );
-  return "disabled";
-}
-
-export function getInstrumentMode(): InstrumentMode {
-  const runnerMode = getCodspeedRunnerMode();
-  // Both "simulation" and "memory" map to "analysis" instrument mode
-  if (runnerMode === "simulation" || runnerMode === "memory") {
-    return "analysis";
-  }
-  return runnerMode; // "disabled" or "walltime"
-}
 
 export const setupCore = () => {
   if (!native_core.isBound) {
@@ -85,7 +48,8 @@ export { optimizeFunction, optimizeFunctionSync } from "./optimization";
 export { wrapWithRootFrame, wrapWithRootFrameSync } from "./rootFrame";
 export * from "./utils";
 export * from "./walltime";
-export type { InstrumentMode };
+export type { InstrumentMode } from "./runnerMode";
+export { getCodspeedRunnerMode, getInstrumentMode } from "./runnerMode";
 export const InstrumentHooks = native_core.InstrumentHooks;
 
 // Marker type constants, sourced from the native addon (which reads them from
