+++ packages/playwright-plugin/package.json
@@ -22,7 +22,7 @@
     "dist"
   ],
   "scripts": {
-    "build": "NODE_NO_WARNINGS=1 rollup -c rollup.config.ts --configPlugin typescript --configImportAttributesKey with",
+    "build": "rollup -c",
     "test": "echo 'no tests'",
     "test/integ": "echo 'no integ tests'",
     "lint": "eslint .",
