+++ packages/tinybench-plugin/package.json
@@ -21,7 +21,7 @@
     "dist"
   ],
   "scripts": {
-    "build": "NODE_NO_WARNINGS=1 rollup -c rollup.config.ts --configPlugin typescript --configImportAttributesKey with",
+    "build": "rollup -c",
     "test": "vitest --run",
     "test/integ": "echo 'no integ tests'",
     "lint": "eslint .",
