+++ packages/vitest-plugin/package.json
@@ -25,7 +25,7 @@
   "homepage": "https://codspeed.io",
   "license": "Apache-2.0",
   "scripts": {
-    "build": "NODE_NO_WARNINGS=1 rollup -c rollup.config.ts --configPlugin typescript --configImportAttributesKey with",
+    "build": "rollup -c",
     "test": "vitest --run",
     "test/integ": "echo 'no integ tests'",
     "lint": "eslint .",
